UPGRADE-TEST: added User tests

+++ reservation/src/main/java/com/upgrade/challenge/reservation/service/impl/ReservationServiceImpl.java
@@ -2,6 +2,7 @@ package com.upgrade.challenge.reservation.service.impl;
 
 import com.upgrade.challenge.reservation.exception.ReservationException;
 import com.upgrade.challenge.reservation.domain.Reservation;
+import com.upgrade.challenge.reservation.exception.UserException;
 import com.upgrade.challenge.reservation.repository.ReservationRepository;
 import com.upgrade.challenge.reservation.service.ReservationService;
 import org.slf4j.Logger;
@@ -12,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
 import java.util.Date;
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -27,48 +29,17 @@ public class ReservationServiceImpl implements ReservationService {
     private ReservationRepository repository;
 
     @Override
-    public Reservation findByStartDate(Date startDate) throws ReservationException {
-        Reservation reservation = null;
-        try {
-            reservation = repository.findByStartDate(startDate);
-        } catch (Exception e) {
-            LOG.error(WARN_MESSAGE, e);
-            throw new ReservationException(WARN_MESSAGE);
-        }
-        return reservation;
-    }
-
-    @Override
-    public Reservation findByEndDate(Date endDate) throws ReservationException {
-        Reservation reservation = null;
-        try {
-            reservation = repository.findByEndDate(endDate);
-        } catch (Exception e) {
-            LOG.error(WARN_MESSAGE, e);
-            throw new ReservationException(WARN_MESSAGE);
-        }
-        return reservation;
-    }
-
-    @Override @Transactional
-    public Reservation save(Reservation reservation) throws ReservationException {
-        Reservation persisted = null;
-        try {
-            persisted = repository.save(reservation);
-        } catch (Exception e) {
-            LOG.error(WARN_MESSAGE, e);
-            throw new ReservationException(WARN_MESSAGE);
-        }
-        return persisted;
-    }
-
-    @Override
     public void cancel(Long id) throws ReservationException {
         try {
-            repository.deleteById(id);
+            if(repository.findById(id).isPresent()) {
+                repository.deleteById(id);
+            }
+            else {
+                throw new ReservationException("There isn't any reservation with id=" + id + ".");
+            }
         } catch (Exception e) {
             LOG.error(WARN_MESSAGE, e);
-            throw new ReservationException(WARN_MESSAGE);
+            throw e;
         }
     }
 
@@ -76,13 +47,19 @@ public class ReservationServiceImpl implements ReservationService {
     public Reservation modify(Reservation reservation) throws ReservationException {
         Reservation modified = null;
         try {
-            Reservation persisted = repository.findById(reservation.getId()).get();
-            persisted.setStartDate(reservation.getStartDate());
-            persisted.setEndDate(reservation.getEndDate());
-            modified = this.save(persisted);
+            List<Reservation> list = repository.findByRange(reservation.getStartDate(), reservation.getEndDate());
+            if(list.isEmpty() || (list.size()==1 && list.get(0).equals(reservation))) {
+                Reservation persisted = list.get(0);
+                persisted.setStartDate(reservation.getStartDate());
+                persisted.setEndDate(reservation.getEndDate());
+                modified = repository.saveAndFlush(persisted);
+            }
+            else {
+                throw new ReservationException("There exists reservations in that period. Please check availability.");
+            }
         } catch(Exception e) {
             LOG.error(WARN_MESSAGE, e);
-            throw new ReservationException(WARN_MESSAGE);
+            throw e;
         }
         return modified;
     }
